rename view directory to panel in shop page
