+++ src/pages/shop/panel/buy/index.tsx
@@ -1,9 +1,9 @@
 import { SelectBox } from '@offer-ui/react'
 import type { ReactElement } from 'react'
 import { useState } from 'react'
-import { LikePanelView } from './panel/like'
-import { OfferPanelView } from './panel/offer'
 import { Styled } from './styled'
+import { LikePanelView } from './view/like'
+import { OfferPanelView } from './view/offer'
 import { useGetMyOffersQuery, useGetLikedPostsQuery } from '@apis'
 import { Tabs } from '@components'
 import { TRADE_ACTIVITY_TYPES, SORT_OPTIONS } from '@constants'
